Conflicts:
	Makefile.am
	configure.ac
	lib/framework/cursors.c
	lib/framework/frame.c
	lib/framework/frameint.h
	lib/framework/input.c
	lib/framework/input.h
	lib/framework/types.h
	lib/framework/utf.h
	lib/gamelib/gtime.c
	lib/ivis_opengl/textdraw.c
	lib/netplay/netplay.c
	lib/netplay/nettypes.h
	lib/widget/editbox.c
	lib/widget/scrap.c
	lib/widget/scrap.h
	macosx/Resources/GenericFramework-Info.plist
	macosx/Warzone.xcodeproj/project.pbxproj
	src/advvis.h
	src/bucket3d.h
	src/combat.h
	src/data.c
	src/displaydef.h
	src/keybind.c
	src/keybind.h
	src/main.cpp
	src/map.h
	src/modding.h
	src/wrappers.c

git-svn-id: https://warzone2100.svn.sourceforge.net/svnroot/warzone2100/branches/qt@9680 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/ivis_opengl/screen.c
@@ -400,7 +400,7 @@ BOOL Init_FBO(unsigned int width, unsigned int height)
 				debug(LOG_ERROR, "Error: FBO Non-framebuffer passed to glCheckFramebufferStatusEXT()!");
 				break;
 			default:
-				debug(LOG_ERROR, "*UNKNOWN FBO ERROR* reported from glCheckFramebufferStatusEXT() for %x!", status);
+				debug(LOG_ERROR, "*UNKNOWN FBO ERROR* reported from glCheckFramebufferStatusEXT() for %x!", (unsigned int)status);
 				break;
 		}
 		FBOinit = false;	//we have a error with the FBO setup
