+++ src/terrain.c
@@ -121,7 +121,7 @@ static bool terrainInitalised = false;
 #define glError() { \
         GLenum err = glGetError(); \
         while (err != GL_NO_ERROR) { \
-                debug(LOG_ERROR, "OpenGL error %i caught at %s:%u\n", err, __FILE__, __LINE__); \
+                debug(LOG_ERROR, "OpenGL error %i caught at %s:%u\n", (int)err, __FILE__, __LINE__); \
                 err = glGetError(); \
         } \
 }
@@ -143,8 +143,8 @@ static void finishDrawRangeElements(void)
 {
 	if (drawRangeElementsStarted && dreCount > 0)
 	{
-		ASSERT(dreEnd - dreStart + 1 <= GLmaxElementsVertices, "too many vertices (%i)", dreEnd - dreStart + 1);
-		ASSERT(dreCount <= GLmaxElementsIndices, "too many indices (%i)", dreCount);
+		ASSERT(dreEnd - dreStart + 1 <= GLmaxElementsVertices, "too many vertices (%i)", (int)(dreEnd - dreStart + 1));
+		ASSERT(dreCount <= GLmaxElementsIndices, "too many indices (%i)", (int)dreCount);
 		glDrawRangeElements(GL_TRIANGLES,
 							dreStart,
 							dreEnd,
@@ -205,8 +205,8 @@ static void addDrawRangeElements(GLenum mode,
 		dreEnd = end;
 	}
 	// make sure we did everything right
-	ASSERT(dreEnd - dreStart + 1 <= GLmaxElementsVertices, "too many vertices (%i)", dreEnd - dreStart + 1);
-	ASSERT(dreCount <= GLmaxElementsIndices, "too many indices (%i)", dreCount);
+	ASSERT(dreEnd - dreStart + 1 <= GLmaxElementsVertices, "too many vertices (%i)", (int)(dreEnd - dreStart + 1));
+	ASSERT(dreCount <= GLmaxElementsIndices, "too many indices (%i)", (int)(dreCount));
 }
 
 /// Get the colour of the terrain tile at the specified position
@@ -706,8 +706,8 @@ bool initTerrain(void)
 	glGetIntegerv(GL_MAX_ELEMENTS_INDICES,  &GLmaxElementsIndices);
 
 	// testing for crappy cards
-	debug(LOG_TERRAIN, "GL_MAX_ELEMENTS_VERTICES: %i", GLmaxElementsVertices);
-	debug(LOG_TERRAIN, "GL_MAX_ELEMENTS_INDICES:  %i", GLmaxElementsIndices);
+	debug(LOG_TERRAIN, "GL_MAX_ELEMENTS_VERTICES: %i", (int)GLmaxElementsVertices);
+	debug(LOG_TERRAIN, "GL_MAX_ELEMENTS_INDICES:  %i", (int)GLmaxElementsIndices);
 	
 	// now we know these values, determine the maximum sector size achievable
 	maxSectorSizeVertices = sqrt(GLmaxElementsVertices/2)-1;
@@ -731,8 +731,8 @@ bool initTerrain(void)
 	{
 		if (sectorSize < 1)
 		{
-			debug(LOG_WARNING, "GL_MAX_ELEMENTS_VERTICES: %i", GLmaxElementsVertices);
-			debug(LOG_WARNING, "GL_MAX_ELEMENTS_INDICES:  %i", GLmaxElementsIndices);
+			debug(LOG_WARNING, "GL_MAX_ELEMENTS_VERTICES: %i", (int)GLmaxElementsVertices);
+			debug(LOG_WARNING, "GL_MAX_ELEMENTS_INDICES:  %i", (int)GLmaxElementsIndices);
 			debug(LOG_WARNING, "maximum sector size due to vertices: %i", maxSectorSizeVertices);
 			debug(LOG_WARNING, "maximum sector size due to indices: %i", maxSectorSizeIndices);
 			debug(LOG_ERROR, "Your graphics card and/or drivers do not seem to support glDrawRangeElements, needed for the terrain renderer.");
